splited the filename functions to Filename.h from FileSystem.h
